feat: implement wrapper around CometBFT RPC client (#1)
* chore: update .gitignore

Update .gitignore to include the vendor/ directory and all subdirectories. It will likely be desirable to check this directory into version control but for now I'm excluding these files for simplicity.

* chore: update README.md

* chore: add basic Makefile

* chore: add .golangci.yml config file

* feat: implement RPC client wrapper around CometBFT's RPC client

Implement a Client type that serves as a wrapper around the CometBFT RPC client. Currently only supports the most recent version of CometBFT but will include support for legacy versions of CometBFT so that we can handle API breaking changes that are introduced upstream.

* chore: update .gitignore

* chore: checkin the vendor/ directory

* feat: bring in slimmed down cometbft client

* chore: check in files missed in previous commit

* remove cometbft dep

* Fix validator updates unmarshal

* chore: update deps

* feat: handle decoding of events from BlockResults & embed the rpc client in our wrapper

Internally try to handle the decoding of events from the BlockResults RPC call internally so that downstream users don't have to consider this detail. We also embed the cometbft rpc client in our wrapper to avoid unnecessary indirection.

* feat: define response types to be used by client wrapper

* refactor: encapsulate comet rpc client & implement methods to invoke rpc calls from wrapper

* test: update test cases

* chore: implement IsOK method for checking query response code

* chore: add Make target for running unit tests

* fix: change json key to camelcase from snake case for proof ops

* fix: allocate memory for slice correctly

* feat: implement missing RPC calls as methods on Client

---------

Co-authored-by: Andrew Gouin <andrew@gouin.io>